refactor(transaction): split module and tighten tx lifecycle
